Add setting to igore filetypes on autocompete
Add setting to .zshrc to ignore files with certain extensions when using
autocomplete after the vim command. This removes a major annoyance of
zsh suggesting files I'd never want to edit such as .class files.

Add setting to .vimrc to do the same this. This is useful when opening
files within vim, such as when splitting the screen.