Disable Terminal Session History Sharing
Keep each terminal's history separate. I don't want cmds I've typed into
a different terminal to come up when scrolling through the history in
a another. If I'm using a terminal for a task, and this is how I use
terminals, this keeps its commands separate and easy to scroll through.